Added some #if around 32bit QuickTime code. 64bit builds without mov support are now possible.

+++ MovFileImport.h
@@ -15,7 +15,7 @@
     NSString        *filePath;
     NSMutableArray  *importCheckArray;
     
-    int chapterTrackId;
+    NSInteger chapterTrackId;
     id  delegate;
 }
